01. Understand the project.
Shall we remember what we learned? In this class we explored the concepts of Tests, we know our Project and we saw the importance of checking the management of state for ease of our tests.

Unit Tests:
Unit tests are the simplest tests to produce! They check small objects like: Classes, methods, variables.

Widget Tests:
Widget tests are slightly more complex and test visible elements in our application that are built on the screen.

Integration Tests:
Integration tests are the most complete and complex tests, which verify interaction with different elements and the entire flow of the application, remembering that these integration tests will not be seen in full in this course.

Project: AluraBank:
A new bank simulation project, with some data on amounts collected and spent, a simple application to learn to test easily.

Importance of Testing States:
When we have a lot of information that transforms our application, we should be smarter, as managing that information complicates test production.