02. Unit Test
What's up? Ready ( o ) to review?! We just saw some useful new concepts for producing quality applications on Flutter:

Folder Test:
There is a folder, in our project, intended only for the tests that we will do in our application. This folder already has a pattern that comes with the initial project. It is essential that tests are created on it.

test ( ):
Function inherited from the Flutter test package, which allows the initiation of a simple unit test. He asks for a description ( that needs to be well structured ) and for a body ( that is responsible for the actions of the test ).

expect ( ):
Method used within the test() to compare the value of the variable and the value expected by that variable. For that, we need to use the actual and the matcher.

group ( ):
Method that groups several tests, to facilitate when we need to run several tests together, facilitating production.