03. Widgets Tests
The time has come to sit down and review what we saw! In this class, we put our hands on the code and built widget tests, we explored several Finders, we learn to create a Matcher and we still saw the difference between automated testing and manual test:

Automated Test Vs Manual Test:
We compared the efficiency of a Manual test ( that is done by the developer, when changing the code and checking if the project actually followed the desired ) and the automated test ( which is done by the framework and is extremely fast and efficient ).

Widget tests:
We learned that Widget tests require a little more care, as they need to be rendered on a emulated screen to check the drawn elements, in addition to having more possibilities of dependencies among other widgets.

pump ( ):
Method capable of waiting for the microprocesses necessary to build / render / draw the elements of the Widgets on the screen. Remember that there are different types of pump() that can better adapt to different tests and specific widgets.

Finders:
Name given to methods capable of searching, on the rendered screen, the desired Widgets, using different search approaches, such as type of Widget, what is your key, among others.

Matcher:
We understand that when we find a widget on the screen, we need to have some degree of control over it. Thus, we can verify how many Widgets are on the screen and which parameters confirm that in fact that Widget is the one sought. Matchers are responsible for this comparison.

Test actions in the code:
When we test our applications, we are faced with actions of interaction with buttons, text boxes, scrollables, among others. Actions are instructions that we give to our tester to interact with certain widgets in order to verify their usability.

We finished Class 3!