04. Mock
Shall we replay what we learn? In this class, we added actions in our test and we produce a Mock with Stub. In addition, we understand more clearly the relations of test dependencies:

Actions:
When we test our applications, we are faced with actions of interaction with buttons, text boxes, scrollables, among others. Actions are instructions that we give to our tester to interact with certain widgets in order to verify their usability.

Using mocks:
We learned that Mock is used to create a double of external dependencies ( like API’s or databases outside the ) project and that it prevents us from having flaky tests.

Operate with Stub:
We found that it is not enough to just create a stuntman, we need this stuntman, in some cases, to give us a return of usable value, and the Stub is made to return this value in the name of the Stuntman.

Understand test dependencies
Now that we have started to go deeper into the tests, we can more easily see how complicated the dependencies between our Classes, methods and Widgets can be, especially in automated tests.

Congratulations! You finished another Flutter course! Now you are one step closer to becoming a professional ( a ) in application production!

Tell us what you think of this course, your feedback is extremely important so that the quality of Flutter courses is always charming.